Replace MCP server and integrate with FastMCP
